Refactoring, Modified GetTransaction to be more similar to bitcoin GetTransaction function

+++ src/BRhodium.Bitcoin.Features.Consensus/Models/TransactionDetailModel.cs
@@ -11,9 +11,9 @@ namespace BRhodium.Bitcoin.Features.Consensus.Models
         public string Account { get; set; }
         public string Address { get; set; }
         public string Category { get; set; }
-        public double Amount { get; set; }
+        public decimal Amount { get; set; }
         [DefaultValue(0D)]
         [JsonProperty(PropertyName = "Fee", DefaultValueHandling = DefaultValueHandling.Ignore)]
-        public double Fee { get; set; }
+        public decimal Fee { get; set; }
     }
 }
